requirements.txt
Flask==3.1.2
Jinja2==3.1.6
neo4j==5.28.2
neomodel==6.0.0
numpy==2.3.5
pandas==2.3.3
python-dotenv==1.2.1
